pyproject.toml
[tool.poetry]
name = "photo-contest-api"
version = "0.1.0"
description = ""
authors = ["JVerissimo <j.verissimo@runtime-revolution.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2"
strawberry-graphql = "^0.171.1"
django-environ = "^0.8.1"
gunicorn = "^20.1.0"
isort = "^5.10.1"
strawberry-graphql-django = "^0.9.4"
black = "^23.3.0"
flake8 = "^6.0.0"
pre-commit = "^3.2.2"
django-pytest = "^0.2.0"
django-debug-toolbar = "^4.0.0"
strawberry-django-plus = "^2.2.0"
pytest = "^7.3.1"
psycopg = "^3.1.10"
factory-boy = "^3.2.1"
psycopg2-binary = "^2.9.6"
pytz = "^2023.3"
django-cors-headers = "^4.0.0"
pytest-django = "^4.5.2"
django-storages = "^1.14"
pillow = "^10.0.1"
boto3 = "^1.28.51"
coverage = "^7.3.1"
django-extensions = "^3.2.3"
djangorestframework = "^3.14.0"
vcrpy = "^5.1.0"
django-allauth = "^0.58.1"
dj-rest-auth = "^5.0.2"
djangorestframework-simplejwt = "^5.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
combine_as_imports = true
